+++ real-estate-domain/src/main/java/io/gig/realestate/domain/realestate/construct/dto/FloorCreateForm.java
@@ -34,11 +34,11 @@ public class FloorCreateForm {
 
     private String companyName;
 
-    private int guaranteePrice;
+    private double guaranteePrice;
 
-    private int rent;
+    private double rent;
 
-    private int management;
+    private double management;
 
     private String termStartDate;
